Fixed the failing pipeline.

+++ commons/src/test/java/commons/NoteTest.java
@@ -79,9 +79,9 @@ public class NoteTest {
 
     @Test
     public void testGetMatchIndexBeginningChar(){
-        ArrayList<Integer> expected = new ArrayList<>();
-        expected.add(0);
-        expected.add(7);
+        ArrayList<Long> expected = new ArrayList<>();
+        expected.add(0L);
+        expected.add(7L);
         assertEquals(expected, note1.getMatchIndices("T"));
     }
     //Todo - add test for full word, and test for searching wrong word.
